refactor: 청결도 컴포넌트 분리
